updated folder structures
