--- Backend/models/Cart.js
@@ -1,19 +1,0 @@
-const mongoose = require("mongoose");
-
-const schema = new mongoose.Schema(
-  {
-    product: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: "Products",
-      required: true
-    },
-
-    quantity: {
-      type: Number,
-      default: 1
-    }
-  },
-  { timestamps: true }
-);
-
-module.exports = mongoose.model("Cart", schema);
