--- Backend/models/User.js
@@ -1,24 +1,0 @@
-const mongoose = require("mongoose");
-
-const schema = new mongoose.Schema(
-  {
-    name: {
-      type: String,
-      required: "name is a required field"
-    },
-
-    email: {
-      type: String,
-      required: true,
-      unique: true
-    },
-
-    password: {
-      type: String,
-      required: true
-    }
-  },
-  { timestamps: true }
-);
-
-module.exports = mongoose.model("user", schema);
