--- README.md
@@ -1,1 +1,0 @@
-# Prerogative-store
