+++ models/Wishlist.js
@@ -1,0 +1,11 @@
+const mongoose = require("mongoose");
+
+const schema = new mongoose.Schema({
+  product: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: "Products",
+    required: true
+  }
+});
+
+module.exports = mongoose.model("Wishlist", schema);
